requirements.txt
streamlit>=1.35
langchain>=0.3
langchain-core>=0.3
langchain-google-genai>=2.0
langchain-chroma>=0.1
langgraph>=0.2
chromadb>=0.5
google-generativeai>=0.8
python-dotenv>=1.0
typing-extensions>=4.9
